requirements.txt
backtrader==1.9.78.123
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
yfinance==0.2.38
joblib==1.4.2
requests==2.32.3
schedule==1.2.1
python-dotenv==1.0.1
transformers==4.41.1
openai==1.30.1
ta==0.11.0